refactor(main): split HexoAgent into domain services
HexoAgent.ts had grown to 1121 lines with 25 public methods, all
sharing the same Hexo instance and a handful of private helpers. This
extracts the implementation into a HexoContext (lifecycle + shared
state/helpers) plus five domain services injected with that context:

- HexoContext: hexo instance, init/settle/ensureReady, updateCache,
  generate, safeResolve, front-matter read/write, static dir checks
- PostService: post queries, document/content read-write, file CRUD
- CategoryService: category listing and bulk re-assign/remove
- TagService: tag listing and detach-from-post
- AssetService: asset listing and deletion
- StatsService: site info, hexo config, aggregate stats

HexoAgent is now a thin facade that delegates to the services, keeping
its public surface (the ISite contract source) byte-for-byte stable, so
the IPC/Web bridge layers need no changes. Behaviour is unchanged; the
existing HexoAgent tests pass with only their white-box injection points
repointed at the internal context.
